+++ Examples/CityTemplate.ps1
@@ -1,0 +1,8 @@
+# City
+Invoke-Generate '[city]' -Count 5         
+
+# City and County
+Invoke-Generate '[city both]' -Count 5    
+
+# County
+Invoke-Generate '[city county]' -Count 5