+++ Examples/CountryTemplate.ps1
@@ -1,0 +1,1 @@
+Invoke-Generate '[country]' -Count 5