+++ Examples/PostalCodeTemplate.ps1
@@ -1,0 +1,1 @@
+Invoke-Generate '[postalcode]' -Count 5