+++ changelog.md
@@ -1,8 +1,9 @@
 # v2.3.3
 
-- Added generators
+- Added generators, see [Examples](./Examples)
     -  `Invoke-Generator '[city]'`
     -  `Invoke-Generator '[company]'`
     -  `Invoke-Generator '[country]'`
     -  `Invoke-Generator '[postalcode]'`
     -  `Invoke-Generator '[email]'`
+